Minimized, divided and optimised (a bit) app with new folders and files close to v0.2
